Working through some LINT things.
BeatService! - Finally worked out the static context/memory leaking issue. A broadcast receiver that is an inner class must be declared as static in order to receive broadcasts when it's defined in the manifest. This places constraints on the methods and variables the receiver can access - they must all also be static. This static context is a no-no. In order to work around it, the receiver needs to be defined as a member level variable, and then the receiver is instantiated in onHandleIntent(); The receiver can receive and process broadcasts. If a context needs to be handed around, it can be passed directly to the method in question.